Fix/separated User api into Resume api
