+++ src/api/User.ts
@@ -94,7 +94,6 @@ export const getFreelancersBySort = async (sortLabel: string) => {
   }
 };
 
-//------------------------------------------------------------------//
 // 마이페이지
 export const getFreelancer = async (userId: string) => {
   let { data: users, error } = await supabase
@@ -131,9 +130,6 @@ export const getFreelancerImage = async (userId: string) => {
   if (error) {
     throw new Error("Error loading images");
   }
-  /**
-   * storage에서 시간순으로 정렬해주는 api를 못찾아서 직접 정렬해줌
-   */
   data.sort((a, b) => {
     const timeA = new Date(a.created_at).getTime();
     const timeB = new Date(b.created_at).getTime();
@@ -155,141 +151,3 @@ export const uploadFreelancerImage = async (userId: string, file: File) => {
 
   return data;
 };
-
-//------------------------------------------------------------------//
-// 이력서 tab
-export const getFreelancerResumeProfileIntro = async (userId: string) => {
-  let { data, error } = await supabase
-    .from("users")
-    .select("resumeProfileIntro")
-    .eq("userId", userId);
-  return data;
-};
-
-export const addFreelancerResumeProfileIntro = async ({
-  profileIntroText,
-  userId /** zustand */,
-  freelancerRole /** zustand */,
-  name /** zustand */,
-  photoURL,
-}: {
-  profileIntroText: string;
-  userId: string;
-  freelancerRole: string;
-  name: string;
-  photoURL: string;
-}) => {
-  const { data, error } = await supabase
-    .from("users")
-    .upsert({
-      userId: userId,
-      role: freelancerRole,
-      resumeProfileIntro: profileIntroText,
-      name: name,
-      photoURL: photoURL,
-    })
-    .select();
-  return data;
-};
-
-export const patchFreelancerResumeProfileIntro = async ({
-  editedProfileIntroText,
-  userId,
-  freelancerRole,
-  name,
-  photoURL,
-}: {
-  editedProfileIntroText: string;
-  userId: string;
-  freelancerRole: string;
-  name: string;
-  photoURL: string;
-}) => {
-  const { data, error } = await supabase
-    .from("users")
-    .upsert([
-      {
-        userId: userId,
-        resumeProfileIntro: editedProfileIntroText,
-        role: freelancerRole,
-        name: name,
-        photoURL: photoURL,
-      },
-    ])
-    .select();
-
-  if (error) {
-    // 오류 처리 (예: 오류 메시지 반환 또는 로깅)
-    throw new Error("Failed to update profile intro");
-  }
-
-  return data;
-};
-
-export const getFreelancerResumeExperience = async (userId: string) => {
-  let { data, error } = await supabase
-    .from("users")
-    .select("resumeExperience")
-    .eq("userId", userId);
-  return data;
-};
-
-export const addFreelancerResumeExperience = async ({
-  newData,
-  userId,
-  freelancerRole,
-  name,
-  photoURL,
-}: {
-  newData: object;
-  userId: string;
-  freelancerRole: string;
-  name: string;
-  photoURL: string;
-}) => {
-  // 기존 데이터 가져오기
-  const { data: existingData, error: existingError } = await supabase // existingData: [{resumeExperience: {…}}]
-    .from("users")
-    .select("resumeExperience")
-    .eq("userId", userId);
-
-  // console.log(existingData);
-
-  // 병합된 데이터
-  if (existingData && Array.isArray(existingData[0]?.resumeExperience)) {
-    // existingData가 존재하고 + 그 안의 resumeExperience가 배열인 경우 (=즉, 기존 데이터가 있을 때)
-    const combinedData = {
-      resumeExperience: [
-        ...(existingData?.[0]?.resumeExperience || []),
-        newData,
-      ],
-    };
-    const { data, error } = await supabase
-      .from("users")
-      .upsert({
-        resumeExperience: combinedData.resumeExperience,
-        userId: userId,
-        role: freelancerRole,
-        name: name,
-        photoURL: photoURL,
-      })
-      .select();
-  } else {
-    // 기존 데이터가 없거나, resumeExperience가 배열이 아닌 경우 예외처리
-    const combinedData = {
-      resumeExperience: [newData],
-    };
-    const { data, error } = await supabase
-      .from("users")
-      .upsert({
-        resumeExperience: combinedData.resumeExperience,
-        userId: userId,
-        role: freelancerRole,
-        name: name,
-        photoURL: photoURL,
-      })
-      .select();
-  }
-};
-
-//------------------------------------------------------------------//
